+++ crates/config/src/lib.rs
@@ -7,14 +7,9 @@ mod validators;
 mod workspace;
 
 pub use errors::{format_error_line, format_figment_errors, ConfigError};
-pub use project::dep::*;
-pub use project::global::GlobalProjectConfig;
-pub use project::task::*;
-pub use project::task_options::*;
 pub use project::*;
 pub use types::*;
 pub use validator::ValidationErrors;
-pub use workspace::node::*;
 pub use workspace::*;
 
 pub fn load_workspace_config_template() -> &'static str {
